Add basic log transmogrifier
This is a rough pass at pulling the events from the database and
serializing them to standard output. Running it with bundle exec
(bare where `mysql2` is installed at the system level) will output all
of the clickthroughs created after the `last_time` value in the
`timestamp.yml` file, then update that value.

The events are logged in JSON format to simplify ingest with Filebeat.

It is not super-well factored at this stage, and there are no tests.
This kind of stuff is hard to test without a bunch of stubs, since it is
basically glue between a few libraries. One improvement would be to
inject both the client and serializer into the query object, but this is
of questionable value for now.

There could also be some checking to verify database credentials, that
the timestamp file is present and writeable, and accepting command
line parameters for filenames to simplify cron/nested shell strings.